% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface-funcs.R
\name{area_link}
\alias{area_link}
\title{Gives the html link of a city or country in the Teleport API}
\usage{
area_link(area, type = "city")
}
\arguments{
\item{area}{Either a city or country name, must be given as a string, the first letter must be capitalized}

\item{type}{"city" by default. A string indicating if the area is a city or country, use "country".}
}
\value{
an html link for the salaries of the given area
}
\description{
Use function city_lookup() to see available cities and country_lookup() to see available countries
}
\examples{
area_link("San Diego")
area_link("United States", "country")

}
