% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max-and-min.R
\name{cities_valid}
\alias{cities_valid}
\title{Checks if Valid Cities in the Teleport API are Given}
\usage{
cities_valid(cities)
}
\arguments{
\item{cities}{A list of city names}
}
\value{
A list of FALSE corresponding to each city given that is not valid
}
\description{
Checks if Valid Cities in the Teleport API are Given
}
