% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spider-chart.R
\name{city_dataframe}
\alias{city_dataframe}
\title{Creates a dataframe of a city's scores out of 10
Use function city_lookup() to see available cities}
\usage{
city_dataframe(city)
}
\arguments{
\item{city}{A city name, must be given as a string, the first letter must be capitalized}
}
\value{
A dataframe of a city's score of all 17 metrics
}
\description{
Creates a dataframe of a city's scores out of 10
Use function city_lookup() to see available cities
}
