% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spider-chart.R
\name{city_link}
\alias{city_link}
\title{Gives the html link of a city in the Teleport API
Use function city_lookup() to see available cities}
\usage{
city_link(city)
}
\arguments{
\item{city}{A city name, must be given as a string, the first letter must be capitalized}
}
\value{
An html link of the city's scores
}
\description{
Gives the html link of a city in the Teleport API
Use function city_lookup() to see available cities
}
