% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max-and-min.R
\name{find_max_find_min}
\alias{find_max_find_min}
\title{Finds the Maximum and the Minimum Ratings in a Dataframe Based on the Teleport API}
\usage{
find_max_find_min(citiesdf, locasecat)
}
\arguments{
\item{citiesdf}{A dataframe consisting of two columns, the first being the city and the second being the category}

\item{locasecat}{A valid category that is in lowercase form}
}
\value{
A list containing the maximum rating with its corresponding city, and the minimum rating with its corresponding city
}
\description{
Finds the Maximum and the Minimum Ratings in a Dataframe Based on the Teleport API
}
