% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max-and-min.R
\name{MaxMinRating}
\alias{MaxMinRating}
\title{Returns the Maximum and Minimum Rating of Cities, Given a Certain Category}
\usage{
MaxMinRating(cities, category)
}
\arguments{
\item{cities}{A list of city names}

\item{category}{A category name; accepts only Education, Housing, or Cost of Living - not case sensitive}
}
\value{
A list containing the maximum rating with its corresponding city, and the minimum rating with its corresponding city
}
\description{
Returns the Maximum and Minimum Rating of Cities, Given a Certain Category
}
