% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface-funcs.R
\name{nearest_city}
\alias{nearest_city}
\title{Gives the closest city and distance to a given location}
\usage{
nearest_city(lat, lon)
}
\arguments{
\item{lat}{the latitude position to search with}

\item{lon}{the longitude position to search with}
}
\value{
A data frame with closest city name and distance(km)
}
\description{
Gives the closest city and distance to a given location
}
\examples{
nearest_city(lat = 37, lon = -120)

}
