% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface-funcs.R
\name{salaries_qt}
\alias{salaries_qt}
\title{Gives a dataframe of the 25th, 50th, and 75th salary percentiles for jobs in a given city}
\usage{
salaries_qt(area, type = "city")
}
\arguments{
\item{area}{Either a city or country name, must be given as a string, the first letter must be capitalized}

\item{type}{"city" by default. A string indicating if the area is a city or country, use "country".}
}
\value{
a dataframe where each row is a job, and the columns are salary percentiles
}
\description{
Gives a dataframe of the 25th, 50th, and 75th salary percentiles for jobs in a given city
}
\examples{
salaries_qt("San Diego")
salaries_qt("United States", "country")

}
