% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spider-chart.R
\name{city_radarchart}
\alias{city_radarchart}
\title{Produces a Spider Chart comparing two cities' metrics from the Teleport API}
\usage{
city_radarchart(city1, city2)
}
\arguments{
\item{city1}{A city name, must be given as a string, the first letter must be capitalized}

\item{city2}{A city name, must be given as a string, the first letter must be capitalized}
}
\description{
Use function city_lookup() to see available cities
}
\examples{
city_radarchart("Winnipeg","Salt Lake City")
city_radarchart("Rome","Glasgow")

}
